2026-01-08/vector-db-similarity-search/requirements.txt
fastapi
uvicorn
chromadb
requests
python-dotenv
pydantic
pytest